requirements.txt
streamlit
scrapy
playwright
langchain
faiss-cpu
openai
python-dotenv